requirements.txt
opencv-python
numpy
sacad
moviepy
yt-dlp 
boto3
tweepy
python-dotenv